chore: apps/web 폴더 구조 설정 (#7)
