+++ apps/web/src/app/layout.tsx
@@ -1,0 +1,29 @@
+import type { Metadata } from 'next'
+import localFont from 'next/font/local'
+import './globals.css'
+
+const geistSans = localFont({
+  src: '../../public/fonts/GeistVF.woff',
+  variable: '--font-geist-sans',
+})
+const geistMono = localFont({
+  src: '../../public/fonts/GeistMonoVF.woff',
+  variable: '--font-geist-mono',
+})
+
+export const metadata: Metadata = {
+  title: 'Create Next App',
+  description: 'Generated by create next app',
+}
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode
+}>) {
+  return (
+    <html lang="en">
+      <body className={`${geistSans.variable} ${geistMono.variable}`}>{children}</body>
+    </html>
+  )
+}
